==> 2_Train_FaceRecognition_with_ML/05_FRM_make_pipeline.py <==
import pandas as pd
import numpy as np
import sklearn
import pickle

import matplotlib.pyplot as plt
import cv2



#Load all models

haar = cv2.CascadeClassifier('./model/haarcascade_frontalface_default.xml') # cascade classifier
model_svm = pickle.load(open('./model/model_svm.pickle',mode='rb')) # machine learning model svm
pca_models = pickle.load(open('./model/pca_dict.pickle',mode='rb')) # pca dictionary

model_pca = pca_models['pca'] # PCA model
mean_face_arr = pca_models['mean_face'] # Mean face



#### Create pipeline --- 11 Steps

# step - 01  -- read image
img = cv2.imread('./test_images/getty_test.jpg') #BGR

# step - 02  -- convert into gray scale
gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)

# step - 03  -- crop the face (haar cascad classifier)
faces = haar.detectMultiScale(gray, 1.5, 3)\

predictions = []

for x,y,w,h in faces:
    roi = gray[y:y+h, x:x+w]
    
    # step - 04 -- Normalization(0-1)
    roi = roi/255.0

    # step - 05 -- resize images (100, 100)
    if(roi.shape[1] > 100):
        roi_resize = cv2.resize(roi, (100, 100), cv2.INTER_AREA)
    else:
        roi_resize = cv2.resize(roi, (100, 100), cv2.INTER_CUBIC)

    # step - 06 -- flattening the image (1, 10000)
    roi_reshape = roi_resize.reshape(1, 10000)
      
    # step - 07 -- subtract with mean
    roi_mean = roi_reshape - mean_face_arr

    # step - 08 -- get eigan image (apply roi mean to pca)
    eigan_image = model_pca.transform(roi_mean)

    # step 09 -- Eigen image for visualization
    eig_img = model_pca.inverse_transform(eigan_image)

    #step 10 -- pass to ml model and get predictions (svm)
    results = model_svm.predict(eigan_image)

    prob_score = model_svm.predict_proba(eigan_image)
    prob_score_max = prob_score.max()

    print(results, prob_score_max)

    # step 11 - generate report
    text = "%s : %d"%(results[0], prob_score_max)

    # defining color based on results
    if results[0]=="male":
        color = (255, 255, 0)
    else:
        color = (255, 0, 255)

    cv2.rectangle(img, (x, y), (x + w, y+h), color, 2)
    cv2.rectangle(img, (x, y-40), (x + w, y), color, -1)
    cv2.putText(img, text, (x,y), cv2.FONT_HERSHEY_PLAIN, 3, (255, 255, 255), 5)
    output = {
        'roi':roi,
        'eig_img':eig_img,
        'prediction_name':results[0],
        'score':prob_score_max
    }

    predictions.append(output)
    print(predictions)


    img_rgb = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
    plt.figure(figsize=(10,10))
    plt.imshow(img_rgb)
    plt.axis('off')
    plt.show()


    ## generate report
for i in range(len(predictions)):
    obj_gray = predictions[i]['roi'] #gray scale
    obj_eig = predictions[i]['eig_img'].reshape(100,100) #eigen image
    plt.subplot(1,2,1)
    plt.imshow(obj_gray,cmap='gray')
    plt.title('Gray ScaleImage')
    plt.axis('off')
    plt.subplot(1,2,2)
    plt.imshow(obj_eig,cmap='gray')
    plt.title('Eigen Image')
    plt.axis('off')
    
    plt.show()
    print('Predicted Gender =',predictions[i]['prediction_name'])
    print('Predicted score = {:,.2f} %'.format(predictions[i]['score']*100))
    
    print('-'*100)
